pyproject.toml
[tool.poetry]
name = "trademaster2-0"
version = "0.1.0"
description = ""
authors = ["sturgis <sturgis.steele@outlook.com>"]
readme = "README.md"
packages = [{include = "trademaster2"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
asyncio = "^3.4.3"
discord-py = "^2.5.2"
python-dotenv = "^1.0.1"
langchain-ollama = "^0.2.3"
langgraph = "^0.3.5"
langsmith = "^0.3.12"
langchain = "^0.3.20"
langchain-community = "^0.3.19"
crewai = "^0.102.0"
langchain-groq = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
